+++ cypress/support/cypress-environment.ts
@@ -3,5 +3,6 @@
  * This is separate from the application environment to allow for different API paths
  */
 export const cypressEnvironment = {
-    apiBasePath: 'api/v1/gateway'
+    apiBasePath: 'api/v1/gateway',
+    geocodeUrl: 'https://api.heigit.org/pelias/v1'
 }